+++ packages/errors-nest/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 0.5.0
+
+### Minor Changes
+
+- Added Health-check related errors
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/errors@0.4.0
+
 ## 0.4.0
 
 ### Minor Changes